@@ -34,79 +34,79 @@
     "Angular 2+"
   ],
   "dependencies": {
-    "@angular/animations": "^7.2.6",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.6",
-    "@angular/compiler": "^7.2.6",
-    "@angular/core": "^7.2.6",
-    "@angular/forms": "^7.2.6",
-    "@angular/http": "^7.2.6",
-    "@angular/platform-browser": "^7.2.6",
-    "@angular/platform-browser-dynamic": "^7.2.6",
-    "@angular/router": "^7.2.6",
-    "@ngx-translate/core": "^10.0.1",
-    "angular-sortablejs": "^2.5.2",
-    "bootstrap": "^4.3.1",
-    "china-division": "^2.1.1",
-    "core-js": "^2.5.4",
-    "highlight.js": "^9.13.1",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/http": "^7.2.16",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@ngx-translate/core": "^11.0.1",
+    "angular-sortablejs": "^2.7.0",
+    "bootstrap": "^4.4.1",
+    "china-division": "^2.2.0",
+    "core-js": "^3.6.4",
+    "highlight.js": "^9.17.1",
     "natives": "^1.1.6",
-    "ng-packagr": "^4.7.1",
-    "ngx-bootstrap": "^4.2.0",
-    "ngx-highlightjs": "^3.0.0",
-    "node-sass": "^4.12.0",
-    "rxjs": "^6.4.0",
-    "sortablejs": "^1.7.0",
-    "tsickle": "^0.34.3",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.29"
+    "ng-packagr": "^5.7.1",
+    "ngx-bootstrap": "^5.3.2",
+    "ngx-highlightjs": "^4.0.2",
+    "node-sass": "^4.13.0",
+    "rxjs": "^6.5.4",
+    "sortablejs": "^1.10.2",
+    "tsickle": "^0.38.0",
+    "tslib": "^1.10.0",
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular-devkit/build-ng-packagr": "~0.13.0",
-    "@angular/cli": "~7.3.2",
-    "@angular/compiler-cli": "^7.2.6",
-    "@angular/language-service": "^7.2.6",
-    "@commitlint/cli": "^7.5.2",
-    "@commitlint/config-conventional": "^7.5.0",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^8.9.5",
-    "@types/webpack": "^4.1.3",
-    "chalk": "^2.4.2",
-    "codelyzer": "~4.5.0",
-    "commander": "^2.19.0",
-    "concurrently": "^3.6.0",
-    "conventional-changelog-cli": "^2.0.12",
+    "@angular-devkit/build-angular": "~0.803.22",
+    "@angular-devkit/build-ng-packagr": "~0.803.22",
+    "@angular/cli": "~8.3.22",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@commitlint/cli": "^8.3.4",
+    "@commitlint/config-conventional": "^8.3.4",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "^13.1.6",
+    "@types/webpack": "^4.41.2",
+    "chalk": "^3.0.0",
+    "codelyzer": "~5.2.1",
+    "commander": "^4.1.0",
+    "concurrently": "^5.0.2",
+    "conventional-changelog-cli": "^2.0.31",
     "cpx": "^1.5.0",
-    "fs-extra": "^8.0.1",
+    "fs-extra": "^8.1.0",
     "gulp": "^4.0.2",
     "gulp-cli": "^2.2.0",
     "gulp-clean-css": "^4.2.0",
-    "gulp-rename": "^1.4.0",
+    "gulp-rename": "^2.0.0",
     "gulp-sass": "^4.0.2",
-    "http-server": "^0.11.1",
-    "husky": "^2.4.1",
-    "jasmine-core": "~2.99.1",
+    "http-server": "^0.12.1",
+    "husky": "^4.0.7",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.1",
+    "karma-jasmine": "~3.1.0",
+    "karma-jasmine-html-reporter": "^1.5.1",
     "ngm-cli": "^1.0.4",
-    "npm-run-all": "^4.1.2",
-    "prettier": "^1.16.4",
-    "pretty-quick": "^1.10.0",
-    "protractor": "~5.3.0",
-    "readline-sync": "^1.4.9",
-    "replace": "^1.1.0",
-    "semver": "^6.0.0",
-    "standard-version": "^5.0.2",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.20.0",
-    "typescript": "~3.2.4",
-    "webpack": "^4.23.1"
+    "npm-run-all": "^4.1.5",
+    "prettier": "^1.19.1",
+    "pretty-quick": "^2.0.1",
+    "protractor": "~5.4.2",
+    "readline-sync": "^1.4.10",
+    "replace": "^1.1.5",
+    "semver": "^7.1.1",
+    "standard-version": "^7.0.1",
+    "ts-node": "~8.6.2",
+    "tslint": "~5.20.1",
+    "typescript": "3.5.2",
+    "webpack": "^4.41.5"
   },
   "ngPackage": {
     "lib": {
