@@ -158,7 +158,7 @@ class SelectionListWithListOptionsComponent {
     `
 })
 class SelectionListWithListOptionsDefaultComponent {
-    @ViewChild(ThySelectionListComponent) thySelectionListComponent: ThySelectionListComponent;
+    @ViewChild(ThySelectionListComponent, { static: true }) thySelectionListComponent: ThySelectionListComponent;
 
     onValueChange(_change: ThySelectionListChange) {}
 }
