@@ -90,7 +90,7 @@ export class ThyFormGroupComponent implements OnInit {
         this._rowFill = inputValueToBoolean(value);
     }
 
-    @ContentChild('formGroup')
+    @ContentChild('formGroup', { static: true })
     public contentTemplateRef: TemplateRef<any>;
 
     constructor(@Optional() private thyParentForm: ThyFormDirective, private thyTranslate: ThyTranslate) {}
